# api/NeoWallet_OpenAPI_v1.yaml
openapi: 3.1.0
info:
  title: NeoWallet API
  version: 1.0.0
  description: NeoWallet MVP API Contract Specification
  contact:
    name: NeoWallet API Team
    email: api@neowallet.com
  license:
    name: Proprietary
    identifier: Proprietary
servers:
  - url: https://api.neowallet.com/api/v1
    description: Production
  - url: https://api-staging.neowallet.com/api/v1
    description: Staging
  - url: https://api-dev.neowallet.com/api/v1
    description: Development
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required:
        - code
        - message
        - requestId
        - timestamp
      properties:
        code:
          type: string
          example: BUDGET_NOT_FOUND
        message:
          type: string
          example: Budget was not found.
        requestId:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
...
    Device:
      type: object
      properties:
        deviceId:
          type: string
          format: uuid
        deviceName:
          type: string
        deviceType:
          type: string
          enum: [ANDROID, IOS, WEB]
        lastActiveAt:
          type: string
          format: date-time
        registeredAt:
          type: string
          format: date-time

    # Family Schemas
    Family:
      type: object
      properties:
        familyId:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 100
        currency:
          type: string
        ownerId:
          type: string
          format: uuid
        memberCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateFamilyRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          maxLength: 100
        currency:
          type: string
    UpdateFamilyRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
        currency:
          type: string
    DeleteFamilyRequest:
      type: object
      required:
        - confirmation
      properties:
        confirmation:
          type: string
    FamilyInvitationRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
        role:
          type: string
          enum: [MEMBER, RESTRICTED]
          default: MEMBER
    FamilyInvitation:
      type: object
      properties:
        invitationId:
          type: string
          format: uuid
        token:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [MEMBER, RESTRICTED]
        expiresAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    FamilyMember:
      type: object
      properties:
        memberId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [OWNER, MEMBER, RESTRICTED]
        joinedAt:
          type: string
          format: date-time
    UpdateFamilyMemberRequest:
      type: object
      required:
        - role
      properties:
        role:
          type: string
          enum: [MEMBER, RESTRICTED]
    AcceptInvitationResponse:
      type: object
      properties:
        familyId:
          type: string
          format: uuid
        memberId:
          type: string
          format: uuid
        role:
          type: string
          enum: [MEMBER, RESTRICTED]
        joinedAt:
          type: string
          format: date-time

    # Financial Overview Schemas
...
paths:
...
  /users/me/devices:
    get:
      summary: Get user's registered devices
      operationId: getUserDevices
      tags:
        - Users
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Devices retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  devices:
                    type: array
                    items:
                      $ref: '#/components/schemas/Device'
        '401':
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Family Paths
  /families:
    post:
      summary: Create a new family
      operationId: createFamily
      tags:
        - Families
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFamilyRequest'
      responses:
        '201':
          description: Family created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Family'
        '401':
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: User already belongs to a family
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /families/{familyId}:
    get:
      summary: Get family details
      operationId: getFamily
      tags:
        - Families
      security:
        - bearerAuth: []
      parameters:
        - name: familyId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Family details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Family'
        '401':
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not a family member
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Family not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      summary: Update family details
      operationId: updateFamily
      tags:
        - Families
      security:
        - bearerAuth: []
      parameters:
        - name: familyId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateFamilyRequest'
      responses:
        '200':
          description: Family details updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Family'
        '401':
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not family owner
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Family not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Delete family
      operationId: deleteFamily
      tags:
        - Families
      security:
        - bearerAuth: []
      parameters:
        - name: familyId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteFamilyRequest'
      responses:
        '204':
          description: Family deleted successfully
        '401':
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not family owner or family has other members
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Family not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /families/{familyId}/invitations:
    post:
      summary: Invite member to family
      operationId: inviteFamilyMember
      tags:
        - Families
      security:
        - bearerAuth: []
      parameters:
        - name: familyId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FamilyInvitationRequest'
      responses:
        '201':
          description: Invitation created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FamilyInvitation'
        '401':
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not family owner or member
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Family not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Email already a member
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /families/{familyId}/members:
    get:
      summary: Get family members
      operationId: getFamilyMembers
      tags:
        - Families
      security:
        - bearerAuth: []
      parameters:
        - name: familyId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Family members retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  members:
                    type: array
                    items:
                      $ref: '#/components/schemas/FamilyMember'
        '401':
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not a family member
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Family not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /families/{familyId}/members/{memberId}:
    put:
      summary: Update family member role
      operationId: updateFamilyMember
      tags:
        - Families
      security:
        - bearerAuth: []
      parameters:
        - name: familyId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: memberId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateFamilyMemberRequest'
      responses:
        '200':
          description: Family member role updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  memberId:
                    type: string
                    format: uuid
                  userId:
                    type: string
                    format: uuid
                  role:
                    type: string
                  updatedAt:
                    type: string
                    format: date-time
        '401':
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not family owner
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Family or member not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Remove family member
      operationId: removeFamilyMember
      tags:
        - Families
      security:
        - bearerAuth: []
      parameters:
        - name: familyId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: memberId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Family member removed successfully
        '401':
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not family owner
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Family or member not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Cannot remove owner
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /family-invitations/{token}/accept:
    post:
      summary: Accept family invitation
      operationId: acceptFamilyInvitation
      tags:
        - Families
      security:
        - bearerAuth: []
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Invitation accepted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptInvitationResponse'
        '401':
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Invitation not found or expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: User already belongs to a family
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /family-invitations/{token}/reject:
    post:
      summary: Reject family invitation
      operationId: rejectFamilyInvitation
      tags:
        - Families
      security:
        - bearerAuth: []
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Invitation rejected successfully
        '401':
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Invitation not found or expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Financial Overview Paths
...